# openapi.yaml
openapi: 3.1.0
info:
  title: PFF Fantasy League API
  version: "1.0.0"
  description: |
    REST interface for the PFF Fantasy League toolkit. All requests must include
    the header `X-API-Key` matching the `PFF_API_KEY` environment variable.
servers:
  - url: https://pff-fantasy-analyzer.onrender.com
    description: Production
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    PlayoffStanding:
      type: object
      properties:
        teamId: {type: string}
        team: {type: string}
        managers: {type: string, nullable: true}
        wins: {type: integer}
        losses: {type: integer}
        ties: {type: integer}
        gamesPlayed: {type: integer}
        gamesRemaining: {type: integer}
        pointsFor: {type: number}
        pointsAgainst: {type: number}
        winPct: {type: number}
        rank: {type: integer}
      required:
        - teamId
        - team
        - wins
        - losses
        - ties
        - gamesPlayed
        - gamesRemaining
        - pointsFor
        - pointsAgainst
        - winPct
        - rank
    PlayoffTeamProjection:
      type: object
      properties:
        teamId: {type: string}
        team: {type: string}
        managers: {type: string, nullable: true}
        wins: {type: integer}
        losses: {type: integer}
        ties: {type: integer}
        gamesPlayed: {type: integer}
        gamesRemaining: {type: integer}
        pointsFor: {type: number}
        pointsAgainst: {type: number}
        winPct: {type: number}
        meanScore: {type: number, nullable: true}
        stdDev: {type: number, nullable: true}
        rating: {type: number, nullable: true}
        ratingZ: {type: number, nullable: true}
        sosRemaining: {type: number, nullable: true}
        benchVolatility: {type: number, nullable: true}
        playoffProbability: {type: number}
        averageSeed: {type: number, nullable: true}
        medianSeed: {type: number, nullable: true}
        bestSeed: {type: integer, nullable: true}
        worstSeed: {type: integer, nullable: true}
      required:
        - teamId
        - team
        - wins
        - losses
        - ties
        - gamesPlayed
        - gamesRemaining
        - pointsFor
        - pointsAgainst
        - winPct
        - playoffProbability
    PlayoffSimulationMeta:
      type: object
      properties:
        runs: {type: integer}
        playoffSpots: {type: integer}
        seed: {type: integer, nullable: true}
        pendingWeeks:
          type: array
          items: {type: integer}
        schedulePath: {type: string}
      required:
        - runs
        - playoffSpots
        - pendingWeeks
        - schedulePath
    PlayoffOddsResponse:
      type: object
      properties:
        standings:
          type: array
          items: {$ref: '#/components/schemas/PlayoffStanding'}
        teams:
          type: array
          items: {$ref: '#/components/schemas/PlayoffTeamProjection'}
        simulation:
          $ref: '#/components/schemas/PlayoffSimulationMeta'
      required:
        - standings
        - teams
        - simulation
    PlayoffTeamDelta:
      type: object
      properties:
        teamId: {type: string}
        team: {type: string}
        playoffProbabilityDelta: {type: number}
        averageSeedDelta: {type: number, nullable: true}
        medianSeedDelta: {type: number, nullable: true}
        bestSeedDelta: {type: number, nullable: true}
        worstSeedDelta: {type: number, nullable: true}
        meanScoreDelta: {type: number, nullable: true}
        stdDevDelta: {type: number, nullable: true}
        ratingDelta: {type: number, nullable: true}
        ratingZDelta: {type: number, nullable: true}
        sosRemainingDelta: {type: number, nullable: true}
        benchVolatilityDelta: {type: number, nullable: true}
      required:
        - teamId
        - team
        - playoffProbabilityDelta
    PlayoffTradeRequest:
      type: object
      properties:
        teamA: {type: string}
        teamB: {type: string}
        sendA:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        sendB:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        simulations:
          type: integer
          minimum: 500
          maximum: 50000
          default: 5000
        playoffTeams:
          type: integer
          minimum: 2
          default: 8
        seed: {type: integer, nullable: true}
        schedulePath: {type: string, nullable: true}
      required:
        - teamA
        - teamB
    PlayoffTradeResponse:
      type: object
      properties:
        baseline: {$ref: '#/components/schemas/PlayoffOddsResponse'}
        scenario: {$ref: '#/components/schemas/PlayoffOddsResponse'}
        delta:
          type: array
          items: {$ref: '#/components/schemas/PlayoffTeamDelta'}
      required: [baseline, scenario, delta]
...
    StarterDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        csvName: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
    BenchDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
        benchScore: {type: number, nullable: true}
        oVAR: {type: number, nullable: true}
    TeamDetail:
      type: object
      properties:
        team: {type: string}
        starters:
          type: array
          items: {$ref: '#/components/schemas/StarterDetail'}
        bench:
          type: array
          items: {$ref: '#/components/schemas/BenchDetail'}
        benchLimit: {type: integer, nullable: true}
...
    ZeroSumEntry:
      type: object
      properties:
        team: {type: string}
        value: {type: number}
        share: {type: number}
        surplus: {type: number}
      required: [team, value, share, surplus]
...
    ZeroSumScarcityMetric:
      type: object
      properties:
        deficit: {type: number}
        pressure: {type: number}
      required: [deficit, pressure]
    ZeroSumConcentrationRisk:
      type: object
      properties:
        starterPositions:
          type: object
          additionalProperties: {type: number}
        benchPositions:
          type: object
          additionalProperties: {type: number}
        slotShares:
          type: object
          additionalProperties: {type: number}
        flexShare: {type: number}
        herfindahl:
          type: object
          properties:
            starters: {type: number}
            bench: {type: number}
            slots: {type: number}
          additionalProperties: false
      required: [starterPositions, benchPositions, slotShares, flexShare, herfindahl]
...
    TeamLeverageResponse:
      type: object
      properties:
        team: {type: string}
        combined: {$ref: '#/components/schemas/ZeroSumEntry'}
        starters: {$ref: '#/components/schemas/ZeroSumEntry'}
        bench: {$ref: '#/components/schemas/ZeroSumEntry'}
        positions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumEntry'}
        benchPositions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumEntry'}
        slots:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumEntry'}
        scarcityPressure:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumScarcityMetric'}
        concentrationRisk:
          $ref: '#/components/schemas/ZeroSumConcentrationRisk'
        leveragePositions:
          type: array
          items: {type: string}
        needPositions:
          type: array
          items: {type: string}
      required: [team, combined, starters, bench, positions, benchPositions, slots, scarcityPressure, concentrationRisk, leveragePositions, needPositions]
...
    TradePiece:
      type: object
      properties:
        group: {type: string}
        name: {type: string}
      required: [group, name]
...
paths:
...
  /playoffs/odds:
    get:
      tags: [playoffs]
      summary: Compute playoff odds for the active league
      operationId: getPlayoffOdds
      x-openai-isConsequential: false
      parameters:
        - name: simulations
          in: query
          description: Number of Monte Carlo simulation runs to execute.
          schema: {type: integer, default: 5000, minimum: 500, maximum: 50000}
        - name: playoffTeams
          in: query
          description: How many teams qualify for the playoffs.
          schema: {type: integer, default: 8, minimum: 2}
        - name: seed
          in: query
          description: Optional RNG seed for deterministic simulation output.
          schema: {type: integer, nullable: true}
        - name: schedulePath
          in: query
          description: Optional override for the schedule CSV path.
          schema: {type: string, nullable: true}
      responses:
        '200':
          description: Playoff probabilities and supporting metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayoffOddsResponse'
  /playoffs/trade:
    post:
      tags: [playoffs]
      summary: Simulate playoff odds before and after a trade
      operationId: simulatePlayoffTrade
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayoffTradeRequest'
            example:
              teamA: JohnBurrows_School
              teamB: KWood_Super_Marios
              sendA:
                - {group: WR, name: George Pickens}
              sendB:
                - {group: WR, name: Drake London}
              simulations: 2000
              playoffTeams: 8
      responses:
        '200':
          description: Baseline and post-trade playoff odds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayoffTradeResponse'
...
  /teams/{team}:
    get:
      tags: [teams]
      summary: Team detail
      operationId: getTeam
      x-openai-isConsequential: false
      parameters:
        - name: team
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Starters/bench detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
  /teams/{team}/leverage:
    get:
      tags: [teams]
      summary: Zero-sum leverage insight for a team
      operationId: getTeamLeverage
      security:
        - ApiKeyAuth: []
      parameters:
        - name: team
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Zero-sum leverage snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamLeverageResponse'
...
security:
  - ApiKeyAuth: []